requirements.txt
appdirs==1.4.3
backports.csv==1.0.7
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configobj==5.0.6
configparser==4.0.2
cssselect==1.1.0
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
funcsigs==1.0.2
future==0.17.1
geographiclib==1.49
geopy==1.20.0
httplib2==0.13.1
idna==2.8
img2pdf==0.3.3
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.4.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
networkx==2.3
neurdflib==5.0.0.post1
nibabel==2.5.0
nipype==1.2.2
numpy==1.17.1
odfpy==1.4.0
openpyxl==2.4.11
packaging==19.2
pandas==0.25.1
parse==1.12.1
pdf2jpg==1.0
Pillow==6.1.0
prov==1.5.3
pydot==1.4.1
pydotplus==2.0.2
pyee==6.0.0
PyMySQL==0.9.3
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
pytz==2019.2
pyxnat==1.2.0.0
PyYAML==5.1.2
rdflib==4.2.2
records==0.5.3
requests==2.22.0
requests-html==0.10.0
schedule==0.6.0
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
tablib==0.13.0
tqdm==4.35.0
traits==5.1.2
urllib3==1.25.3
w3lib==1.21.0
websockets==8.0.2
Werkzeug==0.15.6
wikipedia==1.4.0
xlrd==1.2.0
xlwt==1.3.0